moving itests into main; failing silently
