--- integration-tests/logging/pom.xml
@@ -1,141 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>io.quarkiverse.googlecloudservices</groupId>
-        <artifactId>quarkus-google-cloud-services-integration-tests-parent</artifactId>
-        <version>0.12.0-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>quarkus-google-cloud-services-logging-it</artifactId>
-    <name>Quarkus - Google Cloud Services - Integration Tests - Logging</name>
-
-    <properties>
-        <maven.deploy.skip>true</maven.deploy.skip>
-        <gpg.skip>true</gpg.skip>
-        <native.surefire.skip>${skipTests}</native.surefire.skip>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-resteasy-jsonb</artifactId>
-          </dependency>
-          <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-jackson</artifactId>
-          </dependency>
-
-        <dependency>
-            <groupId>io.quarkiverse.googlecloudservices</groupId>
-            <artifactId>quarkus-google-cloud-logging</artifactId>
-        </dependency>
-
-        <!-- test dependencies -->
-        <dependency>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-junit5</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>io.rest-assured</groupId>
-            <artifactId>rest-assured</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-          <plugin>
-            <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-maven-plugin</artifactId>
-            <version>${quarkus.version}</version>
-            <executions>
-              <execution>
-                <goals>
-                  <goal>build</goal>
-                </goals>
-              </execution>
-            </executions>
-            <configuration>
-              <graalvmHome>${graalvmHome}</graalvmHome>
-            </configuration>
-          </plugin>
-          <plugin>
-            <artifactId>maven-surefire-plugin</artifactId>
-            <configuration>
-              <systemProperties>
-                <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
-              </systemProperties>
-            </configuration>
-          </plugin>
-          <plugin>
-            <groupId>org.apache.maven.plugins</groupId>
-            <artifactId>maven-jar-plugin</artifactId>
-            <executions>
-              <execution>
-                <goals>
-                  <goal>test-jar</goal>
-                </goals>
-              </execution>
-            </executions>
-          </plugin>
-        </plugins>
-      </build>
-      <profiles>
-        <profile>
-          <id>native</id>
-          <activation>
-            <property>
-              <name>native</name>
-            </property>
-          </activation>
-          <build>
-            <plugins>
-              <plugin>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <executions>
-                  <execution>
-                    <goals>
-                      <goal>integration-test</goal>
-                      <goal>verify</goal>
-                    </goals>
-                    <configuration>
-                      <systemProperties>
-                        <native.image.path>${project.build.directory}/${project.build.finalName}-runner</native.image.path>
-                      </systemProperties>
-                    </configuration>
-                  </execution>
-                </executions>
-              </plugin>
-            </plugins>
-          </build>
-          <properties>
-            <quarkus.package.type>native</quarkus.package.type>
-          </properties>
-        </profile>
-        <profile>
-          <id>gcloud</id>
-          <activation>
-              <property>
-                <name>gcloud</name>
-              </property>
-          </activation>
-          <build>
-            <plugins>
-              <plugin>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <configuration>
-                  <systemProperties>
-                    <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
-                    <gcloud.test>true</gcloud.test>
-                  </systemProperties>
-                </configuration>
-              </plugin>
-            </plugins>
-          </build>
-        </profile>
-    </profiles>
-
-</project>
