--- integration-tests/logging/src/main/resources/application.properties
@@ -1,10 +1,0 @@
-quarkus.google.cloud.logging.default-log=test
-quarkus.google.cloud.logging.default-label.app=test
-quarkus.google.cloud.logging.resource.type=generic_node
-quarkus.google.cloud.logging.resource.label.location=eu-north1
-quarkus.google.cloud.logging.resource.label.namespace=test-nodes
-quarkus.google.cloud.logging.resource.label.node_id=192.168.0.2
-quarkus.google.cloud.logging.structured.stack-trace.included=true
-quarkus.google.cloud.logging.structured.stack-trace.rendering=array
-quarkus.google.cloud.logging.gcp-tracing.enabled=false
-quarkus.google.cloud.logging.format=json
