+++ integration-tests/main/pom.xml
@@ -53,6 +53,10 @@
       <groupId>io.quarkiverse.googlecloudservices</groupId>
       <artifactId>quarkus-google-cloud-secret-manager</artifactId>
     </dependency>
+    <dependency>
+      <groupId>io.quarkiverse.googlecloudservices</groupId>
+      <artifactId>quarkus-google-cloud-logging</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>io.quarkus</groupId>
