+++ integration-tests/main/src/main/java/io/quarkiverse/googlecloudservices/it/logging/TestFormatter.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.googlecloudservices.it;
+package io.quarkiverse.googlecloudservices.it.logging;
 
 import java.util.Map;
 import java.util.logging.ErrorManager;
