+++ integration-tests/main/src/main/resources/application.properties
@@ -3,7 +3,7 @@
 #quarkus.google.cloud.project-id=
 
 # We use a dummy test projet id in test for the emulators to work
-%test.quarkus.google.cloud.project-id=test-project
+%test.quarkus.google.cloud.project-id=larsan-net
 %test.quarkus.google.cloud.storage.host-override=http://localhost:8089
 %test.quarkus.google.cloud.firestore.host-override=localhost:8080
 %test.quarkus.google.cloud.spanner.emulator-host=http://localhost:9010
@@ -18,4 +18,14 @@
 # You can load secrets from Google Cloud Secret Manager with the ${sm//<SECRET_ID>} syntax.
 my.database.password=${sm//integration-test}
 
-quarkus.google.cloud.logging.log=test
+## Logging config
+%test.quarkus.google.cloud.logging.default-log=test
+%test.quarkus.google.cloud.logging.default-label.app=test
+%test.quarkus.google.cloud.logging.resource.type=generic_node
+%test.quarkus.google.cloud.logging.resource.label.location=eu-north1
+%test.quarkus.google.cloud.logging.resource.label.namespace=test-nodes
+%test.quarkus.google.cloud.logging.resource.label.node_id=192.168.0.2
+%test.quarkus.google.cloud.logging.structured.stack-trace.included=true
+%test.quarkus.google.cloud.logging.structured.stack-trace.rendering=array
+%test.quarkus.google.cloud.logging.gcp-tracing.enabled=false
+%test.quarkus.google.cloud.logging.format=json
